# [Perf] Display results with 4 significant figures