--- a/.vscode/cspell.json
+++ b/.vscode/cspell.json
@@ -193,6 +193,7 @@
     "ipconfiguration",
     "ipconfigurations",
     "iqmp",
+    "isclass",
     "iscoroutine",
     "iscoroutinefunction",
     "iscsi",