--- a/tools/azure-devtools/src/azure_devtools/perfstress_tests/_perf_stress_runner.py
+++ b/tools/azure-devtools/src/azure_devtools/perfstress_tests/_perf_stress_runner.py
@@ -8,6 +8,7 @@ import asyncio
 import time
 import inspect
 import logging
+import math
 import os
 import pkgutil
 import sys
@@ -207,8 +208,11 @@ class _PerfStressRunner:
             seconds_per_operation = 1 / operations_per_second
             weighted_average_seconds = total_operations / operations_per_second
             self.logger.info(
-                "Completed {:,} operations in a weighted-average of {:,.2f}s ({:,.2f} ops/s, {:,.3f} s/op)".format(
-                    total_operations, weighted_average_seconds, operations_per_second, seconds_per_operation
+                "Completed {:,} operations in a weighted-average of {}s ({} ops/s, {} s/op)".format(
+                    total_operations,
+                    self._format_number(weighted_average_seconds, 4),
+                    self._format_number(operations_per_second, 4),
+                    self._format_number(seconds_per_operation, 4)
                 )
             )
         else:
@@ -226,3 +230,18 @@ class _PerfStressRunner:
 
         self._operation_status_tracker = total_operations
         self.logger.info("{}\t\t{}\t\t{:.2f}".format(current_operations, total_operations, average_operations))
+
+    def _format_number(self, value, min_significant_digits):
+        # Signficant digits are undefined for the number zero, so hardcode to string "0".
+        if value == 0:
+            return "0"
+
+        log = math.log10(abs(value))
+        significant_digits = max(math.ceil(log), min_significant_digits)
+
+        divisor = 10 ** (math.ceil(log) - significant_digits)
+        rounded = divisor * round(value / divisor)
+
+        decimals = max(0, significant_digits - math.floor(log) - 1)
+
+        return ("{:,." + str(decimals) + "f}").format(rounded)