## Before
```
Completed 4,007 operations in a weighted-average of 10.04s (399.13 ops/s, 0.003 s/op)
```

| Test       | Arguments                                                | 12.13.0 | source | %Change |
|------------|----------------------------------------------------------|---------|--------|---------|
| download   | --size 10240 --parallel 64                               | 409.48  | 402.01 | -1.82%  |
| download   | --size 10485760 --parallel 32                            | 45.28   | 43.96  | -2.92%  |
| download   | --size 1073741824 --parallel 1 --warmup 60 --duration 60 | 0.14    | 0.14   | 0.00%   |
| download   | --size 1073741824 --parallel 8 --warmup 60 --duration 60 | 0.39    | 0.39   | 0.00%   |

## After
```
Completed 4,003 operations in a weighted-average of 10.04s (398.8 ops/s, 0.002507 s/op)
```

| Test       | Arguments                                                | 12.13.0 | source  | %Change |
|------------|----------------------------------------------------------|---------|---------|---------|
| download   | --size 10240 --parallel 64                               | 405.4   | 402.3   | -0.76%  |
| download   | --size 10485760 --parallel 32                            | 45.88   | 45.83   | -0.11%  |
| download   | --size 1073741824 --parallel 1 --warmup 60 --duration 60 | 0.1271  | 0.1264  | -0.55%  |
| download   | --size 1073741824 --parallel 8 --warmup 60 --duration 60 | 0.4177  | 0.3913  | -6.32%  |


## Validation Build
https://dev.azure.com/azure-sdk/internal/_build/results?buildId=1743985&view=results